pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arixa"
version = "0.1.0"
description = "Layer 1 deterministic backend for Arixa"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.6.0",
    "pillow>=11.0.0",
    "mediapipe>=0.10.14",
    "opencv-python-headless>=4.10.0",
    "scikit-image>=0.24.0",
    "mahotas>=1.4.15",
    "colour-science>=0.4.6",
    "PyWavelets>=1.7.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
